- Fixed bug whereby the final chunk of data to be written to the db was sometimes mishapen - Greatly simplified PyTablesFrameController._append() code

+++ zounds/analyze/audiostream.py
@@ -26,7 +26,7 @@ class AudioStream(object):
         self.samplerate = samplerate
         self.windowsize = windowsize
         self.stepsize = stepsize
-        self.chunksize = chunksize_seconds * self.samplerate
+        self.chunksize = int(chunksize_seconds * self.samplerate)
         self.done = False
         
     def _read_frames(self,sndfile):
